feat(reports): move reports module files into dedicated directories
This commit restructures the reports module by moving its files into dedicated directories for better organization and maintainability. Specifically:

- Moved `EventsReportFilterDto` to `src/modules/reports/dto/`.
- Moved `ReportsController` to `src/modules/reports/controllers/`.
- Moved `ReportsService` to `src/modules/reports/services/`.

This change improves code clarity and makes it easier to navigate the reports module's files.
