+++ src/modules/reports/reports.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common'
-import { ReportsController } from '../../controllers/reports.controller'
-import { ReportsService } from './reports.service'
+import { ReportsController } from './controllers/reports.controller'
+import { ReportsService } from './services/reports.service'
 import { PrismaService } from '../../services/prisma.service'
 import { LoggerService } from '../../services/logger.service'
 import { MetricsService } from '../metrics/metrics.service'
