+++ src/modules/reports/services/reports.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common'
 import { PrismaClient } from '@prisma/client'
-import { EventsReportFilter } from './events-report-filter.dto'
-import { PrismaService } from '../../services/prisma.service'
-import { LoggerService } from '../../services/logger.service'
-import { MetricsService } from '../metrics/metrics.service'
-import { RevenueReportFilterDto } from '../../reporter/dto/revenue-report-filter.dto'
-import { CorrelationIdService } from '../../services/correlation-id.service'
+import { EventsReportFilter } from '../dto/events-report-filter.dto'
+import { PrismaService } from '../../../services/prisma.service'
+import { LoggerService } from '../../../services/logger.service'
+import { MetricsService } from '../../metrics/metrics.service'
+import { RevenueReportFilterDto } from '../../../reporter/dto/revenue-report-filter.dto'
+import { CorrelationIdService } from '../../../services/correlation-id.service'
 
 @Injectable()
 export class ReportsService {
